requirements.txt
streamlit>=1.28.0
google-genai>=0.2.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0